+++ Assignment2/app.py
@@ -6,11 +6,16 @@ from functools import wraps
 from types import MethodType
 from wsgiref.util import request_uri
 from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity, verify_jwt_in_request, get_jwt
-from flask import Flask, render_template, request, redirect, jsonify, make_response, url_for, redirect
+from flask import Flask, render_template, request, redirect, jsonify, make_response, url_for, redirect, session
 from hashids import Hashids
-
+from flask_cors import CORS, cross_origin
 app = Flask(__name__)
+app.secret_key = "27eduCBA09"
+app.config['SECRET_KEY'] = 'the quick brown fox jumps over the lazy dog'
+app.config['CORS_HEADERS'] = 'Content-Type'
 app.config["JWT_SECRET_KEY"] = "LisimeiAssignmet2"
+cors = CORS(app, resources={r"/*": {"origins": "*"}})
+# CORS(app)
 jwt = JWTManager(app)
 
 hashids = Hashids()
@@ -50,7 +55,7 @@ def login_required():
         def decorator(*args, **kwargs):
             verify_jwt_in_request()
             user = get_jwt()
-            if user['sub'] in USERPWD.keys():
+            if user['sub']:
                 return fn(*args, **kwargs)
             else:
                 return jsonify(msg="Not Registered!"), 403
@@ -59,7 +64,8 @@ def login_required():
     return wrapper
 
 
-@app.route('/<string:hashid>/', methods=('GET', 'POST', 'DELETE', 'PUT'))
+@app.route('/<string:hashid>', methods=('GET', 'POST', 'DELETE', 'PUT'))
+@cross_origin()
 @login_required()
 def handleId(hashid):
     if request.method == 'DELETE':
@@ -82,21 +88,23 @@ def handleId(hashid):
             URLS[hashid] = url
             shorturl = request.host_url + hashid
             print("New shortURL:"+shorturl)
-            return hashid,200
+            return jsonify({'url': URLS[hashid]}), 200
     if request.method == 'GET':
+        print(hashid)
         if hashid in URLS.keys():
-            return redirect(URLS[hashid], 301)
+            print(URLS[hashid])
+            return URLS[hashid]
         else:
             print("no hashid:"+hashid)
             return render_template('404.html'), 404
 
 
 @app.route('/', methods=('GET', 'POST', 'DELETE'))
+@cross_origin()
 @login_required()
 def modifyShortUrl():
     current_user_id = get_jwt_identity()
-    user = USERPWD[current_user_id]
-    print("User:"+user)
+    print("User:" + current_user_id)
     global URL, URLS, shorturl
     if request.method == 'GET':
         return jsonify(URLS), 200
@@ -120,31 +128,3 @@ def modifyShortUrl():
         id = 0
         return render_template('404.html'), 404
 
-def home():
-    return render_template('home.html')
-
-# login url
-@app.route('/users/login', methods=('GET', 'POST'))
-def login():
-    global access_token
-    if request.method == 'POST':
-        uname = request.form['uname']
-        pwd = request.form['psw']
-        if uname in USERPWD.keys() and USERPWD[uname] == pwd:
-            access_token = create_access_token(identity=uname)
-            return jsonify(({"token": access_token})), 200
-        else:
-            return make_response(jsonify({"msg": "FORBIDDEN"}), 403)
-    return render_template('login.html')
-
-
-@ app.route('/users', methods=('GET', 'POST'))
-def register():
-    if request.method == 'POST':
-        uname = request.form['uname']
-        pwd = request.form['psw']
-        pwdc = request.form['pswc']
-        if pwdc == pwd and uname not in USERPWD.keys():
-            USERPWD[uname] = pwd
-        return redirect(url_for('login'))
-    return render_template('register.html')
