+++ Assignment2/app.sh
@@ -1,3 +1,4 @@
 export FLASK_APP=app  
 export FLASK_ENV=development  
-flask run
+flask run --host 0.0.0.0 --port 5002
+
