+++ Assignment2/templates/base.html
@@ -7,7 +7,10 @@
       name="viewport"
       content="width=device-width, initial-scale=1, shrink-to-fit=no"
     />
-
+    <meta http-equiv="Access-Control-Allow-Origin" content="*" />
+    <script>
+      jQuery.support.cors = true;
+    </script>
     <!-- Bootstrap CSS -->
     <link
       rel="stylesheet"
