--- Assignment2/templates/list.html
@@ -1,8 +1,0 @@
-{% extends 'base.html' %} {% block title %}404{% endblock %} {% block content %}
-
-<div class="content">
-  <h2>URL LIST</h2>
-  {{urls}}
-</div>
-
-{% endblock %}
