+++ Assignment2/user.sh
@@ -1,0 +1,3 @@
+export FLASK_APP=user
+export FLASK_ENV=development  
+flask run --host 0.0.0.0 --port 5001
